Don't retry on RemoteLockServices
