+++ atlasdb-config/src/main/java/com/palantir/atlasdb/http/AtlasDbHttpClients.java
@@ -15,10 +15,8 @@
  */
 package com.palantir.atlasdb.http;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.SSLSocketFactory;
 
@@ -26,16 +24,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
-import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.palantir.atlasdb.util.AtlasDbMetrics;
-import com.squareup.okhttp.CipherSuite;
-import com.squareup.okhttp.ConnectionPool;
-import com.squareup.okhttp.ConnectionSpec;
-import com.squareup.okhttp.Interceptor;
-import com.squareup.okhttp.Response;
-import com.squareup.okhttp.TlsVersion;
 
 import feign.Client;
 import feign.Contract;
@@ -47,11 +37,8 @@ import feign.codec.ErrorDecoder;
 import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonEncoder;
 import feign.jaxrs.JAXRSContract;
-import feign.okhttp.OkHttpClient;
 
 public final class AtlasDbHttpClients {
-    private static final int CONNECTION_POOL_SIZE = 100;
-    private static final long KEEP_ALIVE_TIME_MILLIS = TimeUnit.MILLISECONDS.convert(10, TimeUnit.MINUTES);
     private static final int QUICK_FEIGN_TIMEOUT_MILLIS = 1000;
     private static final int QUICK_MAX_BACKOFF_MILLIS = 1000;
     private static final Request.Options DEFAULT_FEIGN_OPTIONS = new Request.Options();
@@ -61,48 +48,6 @@ public final class AtlasDbHttpClients {
     private static final Encoder encoder = new JacksonEncoder(mapper);
     private static final Decoder decoder = new TextDelegateDecoder(new JacksonDecoder(mapper));
     private static final ErrorDecoder errorDecoder = new AtlasDbErrorDecoder();
-
-    private static final ImmutableList<ConnectionSpec> CONNECTION_SPEC_WITH_CYPHER_SUITES = ImmutableList.of(
-            new ConnectionSpec.Builder(ConnectionSpec.MODERN_TLS)
-                    .tlsVersions(TlsVersion.TLS_1_2)
-                    .cipherSuites(
-                            // This GCM cipher suite is for HTTP/2 over TLS1.2 as clients have to
-                            // enable at least one cipher suite not in the blacklist.
-                            // (https://http2.github.io/http2-spec/index.html#BadCipherSuites)
-                            // Timelock server will support HTTP/2 connections, and this will ensure
-                            // all AtlasDB clients have one supported cipher suite.
-                            // See also:
-                            //    - https://http2.github.io/http2-spec/index.html#rfc.section.9.2.2
-                            CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-                            // In an ideal world, we'd use GCM suites, but they're an order of
-                            // magnitude slower than the CBC suites, which have JVM optimizations
-                            // already. We should revisit with JDK9.
-                            // See also:
-                            //  - http://openjdk.java.net/jeps/246
-                            //  - https://bugs.openjdk.java.net/secure/attachment/25422/GCM%20Analysis.pdf
-                            // CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-                            // CipherSuite.TLS_ECDH_RSA_WITH_AES_256_GCM_SHA384,
-                            // CipherSuite.TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256,
-                            // CipherSuite.TLS_RSA_WITH_AES_256_GCM_SHA384,
-                            // CipherSuite.TLS_RSA_WITH_AES_128_GCM_SHA256,
-                            CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384,
-                            CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256,
-                            CipherSuite.TLS_ECDH_RSA_WITH_AES_256_CBC_SHA384,
-                            CipherSuite.TLS_ECDH_RSA_WITH_AES_128_CBC_SHA256,
-                            CipherSuite.TLS_RSA_WITH_AES_128_CBC_SHA256,
-                            CipherSuite.TLS_RSA_WITH_AES_256_CBC_SHA256,
-                            CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-                            CipherSuite.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
-                            CipherSuite.TLS_ECDH_RSA_WITH_AES_256_CBC_SHA,
-                            CipherSuite.TLS_ECDH_RSA_WITH_AES_128_CBC_SHA,
-                            CipherSuite.TLS_RSA_WITH_AES_256_CBC_SHA,
-                            CipherSuite.TLS_RSA_WITH_AES_128_CBC_SHA,
-                            CipherSuite.TLS_EMPTY_RENEGOTIATION_INFO_SCSV)
-                    .build(),
-            ConnectionSpec.CLEARTEXT);
-
-    @VisibleForTesting
-    static final String USER_AGENT_HEADER = "User-Agent";
 
     private AtlasDbHttpClients() {
         // Utility class
@@ -128,7 +73,7 @@ public final class AtlasDbHttpClients {
                         .encoder(encoder)
                         .decoder(decoder)
                         .errorDecoder(errorDecoder)
-                        .client(newOkHttpClient(sslSocketFactory, userAgent))
+                        .client(FeignOkHttpClients.newOkHttpClient(sslSocketFactory, userAgent, type))
                         .target(type, uri),
                 MetricRegistry.name(type, userAgent));
     }
@@ -196,7 +141,8 @@ public final class AtlasDbHttpClients {
             Optional<SSLSocketFactory> sslSocketFactory, Collection<String> endpointUris,
             Request.Options feignOptions, int maxBackoffMillis, Class<T> type, String userAgent) {
         FailoverFeignTarget<T> failoverFeignTarget = new FailoverFeignTarget<>(endpointUris, maxBackoffMillis, type);
-        Client client = failoverFeignTarget.wrapClient(newOkHttpClient(sslSocketFactory, userAgent));
+        Client client = failoverFeignTarget.wrapClient(
+                FeignOkHttpClients.newOkHttpClient(sslSocketFactory, userAgent, type));
         return AtlasDbMetrics.instrument(
                 type,
                 Feign.builder()
@@ -222,37 +168,5 @@ public final class AtlasDbHttpClients {
                 QUICK_MAX_BACKOFF_MILLIS,
                 type,
                 UserAgents.DEFAULT_USER_AGENT);
-    }
-
-    /**
-     * Returns a feign {@link Client} wrapping a {@link com.squareup.okhttp.OkHttpClient} client with optionally
-     * specified {@link SSLSocketFactory}.
-     */
-    private static Client newOkHttpClient(Optional<SSLSocketFactory> sslSocketFactory, String userAgent) {
-        com.squareup.okhttp.OkHttpClient client = new com.squareup.okhttp.OkHttpClient();
-
-        client.setConnectionSpecs(CONNECTION_SPEC_WITH_CYPHER_SUITES);
-        client.setConnectionPool(new ConnectionPool(CONNECTION_POOL_SIZE, KEEP_ALIVE_TIME_MILLIS));
-        client.setSslSocketFactory(sslSocketFactory.orNull());
-        client.interceptors().add(new UserAgentAddingInterceptor(userAgent));
-        return new OkHttpClient(client);
-    }
-
-    private static final class UserAgentAddingInterceptor implements Interceptor {
-        private final String userAgent;
-
-        private UserAgentAddingInterceptor(String userAgent) {
-            Preconditions.checkNotNull(userAgent, "User Agent should never be null.");
-            this.userAgent = userAgent;
-        }
-
-        @Override
-        public Response intercept(Chain chain) throws IOException {
-            com.squareup.okhttp.Request requestWithUserAgent = chain.request()
-                    .newBuilder()
-                    .addHeader(USER_AGENT_HEADER, userAgent)
-                    .build();
-            return chain.proceed(requestWithUserAgent);
-        }
     }
 }
